Added the notebook for local linear analysis of LH instabilities in a MN
First Notebook of the repository added. The notebook computes the gyroviscous tensor divergence at the first order, and carries out the derivation of the electron response to fluctuations in the electrostatic potential. It serves as a symbolic and algebraic support for the computations carried out in the manuscript Local Analysis of Lower-Hybrid Instabilities in a magnetic nozzle, by M. Ripoli, M. Merino and E. Ahedo.